Add cvs Id: tag.  Quell compiler warning about signed/unsigned  comparison.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/branches/mime@1146 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ bogofilter/tests/deqp.c
@@ -1,3 +1,5 @@
+/* $Id$ */
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -10,7 +12,7 @@ static void die(void) {
 }
 
 int main(void) {
-    long size;
+    size_t size;
     char *buf;
 
     if (fseek(stdin, 0, SEEK_END)) die();
